+++ src/main/java/com/example/demo/repository/DiseaseHistoryRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface DiseaseHistoryRepository extends JpaRepository<DiseaseHistory, Integer> {
-    Optional<DiseaseHistory> findFirstByClientIdAndDisease(Integer clientId, String disease);
+public interface DiseaseHistoryRepository extends JpaRepository<DiseaseHistory, Long> {
+    Optional<DiseaseHistory> findFirstByClientIdAndDisease(Long clientId, String disease);
     List<DiseaseHistory> findByClientId(Long clientId);
     List<DiseaseHistory> findByDoctorId(Long doctorId);
